+++ HISTORY.md
@@ -4,6 +4,7 @@
 
 ### 2025-01-24
 * blocks, objects, threads: special cased RANDOM reporter's input slots to evaluate as String instead of Number, hopefully fixes some extension issues, e.g. Hummingbird
+* new dev version for testing
 
 ## 10.4.2:
 * **Notable Fixes:**